test: split handlers and payload tests
