+++ handler/payload/posts_response_test.go
@@ -1,28 +1,11 @@
 package payload
 
 import (
-	"net/http/httptest"
 	"testing"
 	"time"
 
 	"github.com/oullin/database"
 )
-
-func TestGetPostsFiltersFrom(t *testing.T) {
-	req := IndexRequestBody{Title: "t", Author: "a", Category: "c", Tag: "g", Text: "x"}
-	f := GetPostsFiltersFrom(req)
-	if f.Title != "t" || f.Author != "a" || f.Category != "c" || f.Tag != "g" || f.Text != "x" {
-		t.Fatalf("unexpected filters: %+v", f)
-	}
-}
-
-func TestGetSlugFrom(t *testing.T) {
-	r := httptest.NewRequest("GET", "/posts/s", nil)
-	r.SetPathValue("slug", "  SLUG ")
-	if s := GetSlugFrom(r); s != "slug" {
-		t.Fatalf("slug %s", s)
-	}
-}
 
 func TestGetPostsResponse(t *testing.T) {
 	now := time.Now()
